ch4/ofi: Avoid warnings for unsigned comparisons
Without casting the uint64_t to a signed integer, we see warnings like this:

src/mpid/ch4/netmod/ofi/ofi_win.c: In function ‘win_allgather’:
src/mpid/ch4/netmod/ofi/ofi_win.c:145:31: warning: comparison of unsigned expression < 0 is always false [-Wtype-limits]
  145 |             if (local_key + 1 < MPIDI_OFI_NUM_OPTIMIZED_MEMORY_REGIONS) {
      |                               ^
src/mpid/ch4/netmod/ofi/ofi_win.c:152:27: warning: comparison of unsigned expression >= 0 is always true [-Wtype-limits]
  152 |         if (local_key + 1 >= MPIDI_OFI_NUM_OPTIMIZED_MEMORY_REGIONS ||
      |                           ^~

Casting this value to an int64_t loses a bit, but I don't think we
expect to use the highest order bit any time soon.

+++ src/mpid/ch4/netmod/ofi/ofi_win.c
@@ -140,14 +140,14 @@ static int win_allgather(MPIR_Win * win, void *base, int disp_unit)
                                        MPI_UNSIGNED, MPI_MAX, comm_ptr, &errflag);
             MPIR_ERR_CHECK(mpi_errno);
 
-            if (local_key + 1 < MPIDI_OFI_NUM_OPTIMIZED_MEMORY_REGIONS) {
+            if ((int64_t) local_key + 1 < MPIDI_OFI_NUM_OPTIMIZED_MEMORY_REGIONS) {
                 MPIDI_OFI_global.global_max_optimized_mr_key = local_key + 1;
                 MPIDI_OFI_WIN(win).mr_key = local_key;
             }
         }
         /* Assign regular memory registration key if the optimized one is
          * not requested or exhausted */
-        if (local_key + 1 >= MPIDI_OFI_NUM_OPTIMIZED_MEMORY_REGIONS ||
+        if ((int64_t) local_key + 1 >= MPIDI_OFI_NUM_OPTIMIZED_MEMORY_REGIONS ||
             !MPIDIG_WIN(win, info_args).optimized_mr) {
             /* Makes sure that regular mr key does not fall within optimized mr key range */
             MPIDI_OFI_WIN(win).mr_key =
